Department Controller GetEmployeesByDepartmentId Method Polls were Made

+++ EndProject.Domain/Employee.cs
@@ -5,7 +5,7 @@ namespace EndProject.Domain
     public class Employee:BaseEntity
     {
         public string SurName { get; set; }
-        public int Age { get; set; }
+        public byte Age { get; set; }
         public string Adress { get; set; }
         public Department Department { get; set; }
         public override string ToString()
